pixel: refix 16bpp builds

+++ source/common/pixel.cpp
@@ -303,7 +303,7 @@ inline int _sa8d_8x8(pixel *pix1, intptr_t i_pix1, pixel *pix2, intptr_t i_pix2)
         sum += (sum_t)b0 + (b0 >> BITS_PER_SUM);
     }
 
-    return sum;
+    return (int)sum;
 }
 
 int CDECL sa8d_8x8(pixel *pix1, intptr_t i_pix1, pixel *pix2, intptr_t i_pix2)
